+++ opentelemetry-etw-logs/CHANGELOG.md
@@ -24,6 +24,8 @@
 
 - Added support for populating cloud `role` and `roleInstance` from Resource's `service.name` and `service.instance.id` attributes respectively.
 
+- Exporter now unregister the Etw provider on `shutdown()`.
+
 ## v0.7.0
 
 - Bump msrv to 1.75.0